group `ty::coherence` in one file
