--- compiler/rustc_middle/src/ty/impl_subject.rs
@@ -1,7 +1,0 @@
-use crate::ty::{TraitRef, Ty};
-
-#[derive(Copy, Clone, PartialEq, Eq, Debug, TypeFoldable, TypeVisitable)]
-pub enum ImplSubject<'tcx> {
-    Trait(TraitRef<'tcx>),
-    Inherent(Ty<'tcx>),
-}
